+++ ocaml_version.ml
@@ -42,7 +42,7 @@ let parse s =
     end
 
 let of_string s =
-  try Result.Ok (parse s) with
+  try Ok (parse s) with
   | _ -> Error (`Msg (Printf.sprintf "Unable to parse OCaml version '%s'" s))
 
 let of_string_exn s =
@@ -167,24 +167,54 @@ let string_of_arch = function
   | `I386 -> "i386"
 
 let arch_of_string = function
-  | "arm64" | "aarch64" -> Result.Ok `Aarch64
-  | "amd64" | "x86_64" -> Result.Ok `X86_64
-  | "i386"  | "i686" | "686" | "386" -> Result.Ok `I386
-  | "arm32" | "arm32v7" | "aarch32" -> Result.Ok `Aarch32
-  | "ppc64le" -> Result.Ok `Ppc64le
-  | arch -> Result.Error (`Msg ("Unknown architecture " ^ arch))
+  | "arm64" | "aarch64" -> Ok `Aarch64
+  | "amd64" | "x86_64" -> Ok `X86_64
+  | "i386"  | "i686" | "686" | "386" -> Ok `I386
+  | "arm32" | "arm32v7" | "aarch32" -> Ok `Aarch32
+  | "ppc64le" -> Ok `Ppc64le
+  | arch -> Error (`Msg ("Unknown architecture " ^ arch))
 
 let arch_of_string_exn a =
   match arch_of_string a with
-  | Result.Ok a -> a
-  | Result.Error (`Msg m) -> raise (Invalid_argument m)
+  | Ok a -> a
+  | Error (`Msg m) -> raise (Invalid_argument m)
+
+let to_opam_arch = function
+  | `I386 -> "x86_32"
+  | `X86_64 -> "x86_64"
+  | `Ppc64le -> "ppc64"
+  | `Aarch32 -> "arm32"
+  | `Aarch64 -> "arm64"
+
+let of_opam_arch = function
+  | "x86_32" -> Some `I386
+  | "x86_64" -> Some `X86_64
+  | "ppc64" -> Some `Ppc64le
+  | "arm32" -> Some `Aarch32
+  | "arm64" -> Some `Aarch64
+  | _ -> None
+
+let to_docker_arch = function
+   | `I386 -> "386"
+   | `X86_64 -> "amd64"
+   | `Ppc64le -> "ppc64le"
+   | `Aarch32 -> "arm"
+   | `Aarch64 -> "arm64"
+
+let of_docker_arch = function
+  | "386" -> Some `I386
+  | "amd64" -> Some `X86_64
+  | "ppc64le" -> Some `Ppc64le
+  | "arm" -> Some `Aarch32
+  | "arm64" -> Some `Aarch64
+  | _ -> None
 
 module Since = struct
   let bytes = Releases.v4_03_0
 
   let arch (a:arch) =
     match a with
-    | `I386 -> Releases.v4_06_0 (* TODO can be ealier *)
+    | `I386 -> Releases.v4_06_0 (* TODO can be earlier *)
     | `Aarch32 -> Releases.v4_06_0
     | `Aarch64 -> Releases.v4_05_0
     | `Ppc64le -> Releases.v4_06_0
@@ -205,15 +235,34 @@ module Has = struct
 end
 
 module Configure_options = struct
-  type o = [ `Afl | `Flambda | `Default_unsafe_string | `Force_safe_string | `Frame_pointer ]
+  type o = [
+      `Afl
+    | `Default_unsafe_string
+    | `Disable_flat_float_array
+    | `Flambda
+    | `Force_safe_string
+    | `Frame_pointer
+    | `No_naked_pointers ]
 
-  let to_description t =
+  let compare a b =
+    (* For backwards compat reasons, fp always comes first. *)
+    match a,b with
+    | `Frame_pointer, `Frame_pointer -> 0
+    | `Frame_pointer, _ -> (-1)
+    | _, `Frame_pointer -> 1
+    | a, b -> Stdlib.compare a b
+
+  let equal a b = compare a b = 0
+
+  let to_description (t:o) =
     match t with
     | `Afl -> "AFL (fuzzing) support"
     | `Flambda -> "flambda inlining"
     | `Default_unsafe_string -> "default to unsafe strings"
     | `Force_safe_string -> "force safe string mode"
     | `Frame_pointer -> "frame pointer"
+    | `No_naked_pointers -> "forbid unboxed pointers"
+    | `Disable_flat_float_array -> "disable float array unboxing"
 
   let to_string t =
     match t with
@@ -222,33 +271,56 @@ module Configure_options = struct
     | `Default_unsafe_string -> "default-unsafe-string"
     | `Force_safe_string -> "force-safe-string"
     | `Frame_pointer -> "fp"
+    | `No_naked_pointers -> "nnp"
+    | `Disable_flat_float_array -> "no-flat-float-array"
 
-  let to_configure_flag t =
-    match t with
-    | `Afl -> "-afl-instrument"
-    | `Flambda -> "-flambda"
-    | `Default_unsafe_string -> "-default-unsafe-string"
-    | `Force_safe_string -> "-force-safe-string"
-    | `Frame_pointer -> "-with-frame-pointer"
+  let of_string = function
+    | "afl" -> Some `Afl
+    | "flambda" -> Some `Flambda
+    | "default-unsafe-string" -> Some `Default_unsafe_string
+    | "force-safe-string" -> Some `Force_safe_string
+    | "fp" -> Some `Frame_pointer
+    | "nnp" -> Some `No_naked_pointers
+    | "no-flat-float-array" -> Some `Disable_flat_float_array
+    | _ -> None
 
+  let to_t t = function
+    | [] -> with_variant t None
+    | ol ->
+      List.sort compare ol |> List.map to_string |> String.concat "+" |>
+      fun s -> with_variant t (Some s)
+
+  let of_t t =
+    match t.extra with None -> Ok [] | Some extra ->
+    String.split_on_char '+' extra |>
+    List.map (fun b -> match of_string b with
+      | None -> Error (`Msg ("unknown variant: " ^ b))
+      | Some v -> Ok v) |>
+    List.fold_left (fun a b ->
+      match a, b with
+      | Ok a, Ok b -> Ok (List.sort compare (b :: a))
+      | _, Error b -> Error b
+      | Error a, _-> Error a) (Ok [])
+
+  let to_configure_flag {major; minor; _} o =
+    if major < 5 && minor < 8 then (* pre autoconf *)
+      match o with
+      | `Afl -> "-afl-instrument"
+      | `Flambda -> "-flambda"
+      | `Default_unsafe_string -> "-default-unsafe-string"
+      | `Force_safe_string -> "-force-safe-string"
+      | `Frame_pointer -> "-with-frame-pointer"
+      | _ -> ""
+    else (* post autoconf *)
+      match o with
+      | `Afl -> "--with-afl"
+      | `Flambda -> "--enable-flambda"
+      | `Default_unsafe_string -> "--enable-default-unsafe-string"
+      | `Force_safe_string -> "--force-safe-string"
+      | `Frame_pointer -> "--enable-frame-pointers"
+      | `No_naked_pointers -> "--disable-naked-pointers"
+      | `Disable_flat_float_array -> "--disable-flat-float-array"
 end
-
-let compiler_variants arch {major; minor; _} =
-    match major,minor,arch with
-    | 4,12,`X86_64 -> [[]; [`Afl]; [`Flambda]]
-    | 4,11,`X86_64 -> [[]; [`Afl]; [`Flambda]]
-    | 4,10,`X86_64 -> [[]; [`Afl]; [`Flambda]]
-    | 4,9,`X86_64 -> [[]; [`Afl]; [`Flambda]; [`Frame_pointer]; [`Frame_pointer;`Flambda]; [`Default_unsafe_string]]
-    | 4,8,`X86_64 -> [[]; [`Afl]; [`Flambda]; [`Frame_pointer]; [`Frame_pointer;`Flambda]; [`Default_unsafe_string]; [`Force_safe_string]]
-    | 4,10,_ -> [[]; [`Afl]; [`Flambda]]
-    | 4,9,_ -> [[]; [`Afl]; [`Flambda]; [`Default_unsafe_string]]
-    | 4,8,_ -> [[]; [`Afl]; [`Flambda]; [`Default_unsafe_string]; [`Force_safe_string]]
-    | 4,7,_ -> [[]; [`Afl]; [`Flambda]; [`Default_unsafe_string]; [`Force_safe_string]]
-    | 4,6,_ -> [[]; [`Afl]; [`Flambda]; [`Default_unsafe_string]; [`Force_safe_string]]
-    | 4,5,_ -> [[]; [`Afl]; [`Flambda]]
-    | 4,4,_ -> [[]; [`Flambda]]
-    | 4,3,_ -> [[]; [`Flambda]]
-    | _ -> [[]]
 
 module Sources = struct
   let trunk = Releases.v4_12
@@ -258,6 +330,34 @@ module Sources = struct
     | major, minor, _ when major = trunk.major && minor = trunk.minor -> "trunk"
     | _ -> to_string (with_variant ov None)
 end
+
+let trunk_variants (arch:arch) =
+  let base = [[`No_naked_pointers]; [`Afl]; [`Flambda]; [`Disable_flat_float_array]] in
+  let arch_opts =
+    match arch with
+    |`X86_64 -> [[`Frame_pointer]; [`Frame_pointer;`Flambda]]
+    |_ -> []
+  in
+  List.map (Configure_options.to_t Sources.trunk) (base @ arch_opts)
+
+let compiler_variants arch ({major; minor; _} as t) =
+  let f = List.map (Configure_options.to_t t) in
+  match major,minor,arch with
+    | 4,12,arch  -> trunk_variants arch
+    | 4,11,`X86_64 -> f [[]; [`Afl]; [`Flambda]; [`Frame_pointer]]
+    | 4,10,`X86_64 -> f [[]; [`Afl]; [`Flambda]; [`Frame_pointer]]
+    | 4,9,`X86_64 -> f [[]; [`Afl]; [`Flambda]; [`Frame_pointer]]
+    | 4,8,`X86_64 -> f [[]; [`Afl]; [`Flambda]; [`Frame_pointer]]
+    | 4,11,_ -> f [[]; [`Afl]; [`Flambda]]
+    | 4,10,_ -> f [[]; [`Afl]; [`Flambda]]
+    | 4,9,_ -> f [[]; [`Afl]; [`Flambda]]
+    | 4,8,_ -> f [[]; [`Afl]; [`Flambda]]
+    | 4,7,_ -> f [[]; [`Afl]; [`Flambda]]
+    | 4,6,_ -> f [[]; [`Afl]; [`Flambda]]
+    | 4,5,_ -> f [[]; [`Afl]; [`Flambda]]
+    | 4,4,_ -> f [[]; [`Flambda]]
+    | 4,3,_ -> f [[]; [`Flambda]]
+    | _ -> f [[]]
 
 module Opam = struct
 
@@ -272,12 +372,9 @@ module Opam = struct
     let variant_switch t vs =
       match vs with
       | [] -> with_variant t None
-      | vs ->
-        let v = String.concat "+" (List.map Configure_options.to_string vs) in
-        with_variant t (Some v)
+      | vs -> Configure_options.to_t t vs
 
     let switches arch t =
-      compiler_variants arch t |>
-      List.map (fun vs -> variant_switch t vs)
+      compiler_variants arch t
   end
 end