+++ ocaml_version.mli
@@ -40,7 +40,7 @@ val to_string : ?sep:char -> t -> string
     Docker container tags from OCaml version strings, where
     only dashes and alphanumeric characters are allowed. *)
 
-val of_string : string -> (t, [> `Msg of string ]) Result.result
+val of_string : string -> (t, [> `Msg of string ]) result
 (** [of_string t] will parse the version string in [t].
     The return value is compatible with the {!Result}
     combinators defined in the [rresult] library. *)
@@ -84,7 +84,7 @@ val string_of_arch : arch -> string
     {{:https://golang.org/doc/install/source#environment}GOARCH}
     convention used by Golang. *)
 
-val arch_of_string : string -> (arch, [> `Msg of string ]) Result.result
+val arch_of_string : string -> (arch, [> `Msg of string ]) result
 (** [arch_of_string t] will parse the architecture string in [t].
     The return value is compatible with the {!Result}
     combinators defined in the [rresult] library. This function
@@ -99,6 +99,23 @@ val arch_of_string_exn: string -> arch
 val arch_is_32bit: arch -> bool
 (** [arch_is_32bit t] will return [true] if the architecture has
      a 32-bit wordsize. *)
+
+val to_opam_arch : arch -> string
+(** [to_opam_arch arch] will return a string that is compatible
+    with opam's [%{arch}%] variable. *)
+
+val of_opam_arch : string -> arch option
+(** [of_opam_arch s] will try to convert [s] that represents an
+    opam [%{arch}%] variable to an {!arch} value. *)
+
+val to_docker_arch : arch -> string
+(** [to_docker_arch arch] will return a string that is compatible
+    with a Docker multiarch property.  This can be used in
+    [--platform] flags to [docker run], for example. *)
+
+val of_docker_arch : string -> arch option
+(** [of_docker_arch s] will try to convert [s] that represents
+    a Docker multiarch variable to an {!arch} value. *)
 
 (** {2 Accessors} *)
 
@@ -334,24 +351,50 @@ module Configure_options : sig
   type o =
     [ `Afl
     | `Default_unsafe_string
+    | `Disable_flat_float_array
     | `Flambda
     | `Force_safe_string
-    | `Frame_pointer ]
+    | `Frame_pointer
+    | `No_naked_pointers ]
+
   (** Configuration options available at compiler build time. *)
 
   val to_string : o -> string
   (** [to_string o] returns a compact representation of {!o} suitable for use in opam version strings. *)
 
+  val of_string : string -> o option
+  (** [of_string s] will parse the output of {!to_string} back into an option {!o}. Returns {!None} if
+      the string input is unknown. *)
+
+  val of_t : t -> (o list, [> `Msg of string ]) result
+  (** [of_t t] will parse the [extra] field of [t] and return a list of configure options that it
+      represents. Unknown options in the extra field will result in an [Error] being returned. *)
+
+  val to_t : t -> o list -> t
+  (* [to_t t ol] will replace the [extra] field of [t] with the list of options represented in [ol]. *)
+
   val to_description : o -> string
   (** [to_description o] returns a human-readable representation of {!o}. *)
 
-  val to_configure_flag : o -> string
+  val to_configure_flag : t -> o -> string
   (** [to_configure_flag o] returns a string that can be passed to OCaml's [configure] script to activate that feature. *)
+
+  val compare : o -> o -> int
+  (** [compare a b] will return -1 if [a] is < [b], 0 if they are equal, or 1 if [a] > [b]. For backwards
+      compatibility reasons, {!`Frame_pointer} always comes first in comparisons. *)
+
+  val equal : o -> o -> bool
+  (** [equal a b] will return {!true} if [a=b] *)
+
 end
 
-val compiler_variants : arch -> t -> Configure_options.o list list
+val compiler_variants : arch -> t -> t list
 (** [compiler_variants v] returns a list of configuration options that are available and useful
     for version [v] of the compiler. *)
+
+val trunk_variants : arch -> t list
+(** [trunk_variants v] returns a list of OCaml version configurations that should be working and tested
+    on the trunk version of the compiler. *)
 
 (** Opam compiler switches.
     These are available from the public {{:https://github.com/ocaml/opam-repository}opam-repository}. *)